requirements/requirements.txt
Django==3.1.2
psycopg2==2.8.3
django-environ==0.4.5
factory-boy==2.12.0
faker==4.1.1 
flake8==3.8.3
codecov==2.1.8
coverage==5.2.1 
pytest-cov==2.10.0   
pytest-django==3.9.0 
pytest==6.0.1
attrs==19.3.0
idna==2.10
iniconfig==1.0.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
toml==0.10.1
